# Speed up function `compose_transformations` by 24%