--- a/kornia/geometry/linalg.py
+++ b/kornia/geometry/linalg.py
@@ -37,11 +37,11 @@ __all__ = [
 
 
 def compose_transformations(trans_01: Tensor, trans_12: Tensor) -> Tensor:
-    r"""Compose two homogeneous transformations.
+    """Compose two homogeneous transformations.
 
     .. math::
         T_0^{2} = \begin{bmatrix} R_0^1 R_1^{2} & R_0^{1} t_1^{2} + t_0^{1} \\
-        \mathbf{0} & 1\end{bmatrix}
+        \\mathbf{0} & 1\\end{bmatrix}
 
     Args:
         trans_01: tensor with the homogeneous transformation from
@@ -69,24 +69,24 @@ def compose_transformations(trans_01: Tensor, trans_12: Tensor) -> Tensor:
     if not ((trans_12.dim() in (2, 3)) and (trans_12.shape[-2:] == (4, 4))):
         raise ValueError(f"Input trans_12 must be a of the shape Nx4x4 or 4x4. Got {trans_12.shape}")
 
-    if not trans_01.dim() == trans_12.dim():
+    if trans_01.dim() != trans_12.dim():
         raise ValueError(f"Input number of dims must match. Got {trans_01.dim()} and {trans_12.dim()}")
 
     # unpack input data
-    rmat_01: Tensor = trans_01[..., :3, :3]  # Nx3x3
-    rmat_12: Tensor = trans_12[..., :3, :3]  # Nx3x3
-    tvec_01: Tensor = trans_01[..., :3, -1:]  # Nx3x1
-    tvec_12: Tensor = trans_12[..., :3, -1:]  # Nx3x1
+    rmat_01 = trans_01[..., :3, :3]
+    rmat_12 = trans_12[..., :3, :3]
+    tvec_01 = trans_01[..., :3, 3:]
+    tvec_12 = trans_12[..., :3, 3:]
 
     # compute the actual transforms composition
-    rmat_02: Tensor = torch.matmul(rmat_01, rmat_12)
-    tvec_02: Tensor = torch.matmul(rmat_01, tvec_12) + tvec_01
+    rmat_02 = torch.matmul(rmat_01, rmat_12)
+    tvec_02 = torch.matmul(rmat_01, tvec_12) + tvec_01
 
-    # pack output tensor
-    trans_02: Tensor = zeros_like(trans_01)
-    trans_02[..., :3, 0:3] += rmat_02
-    trans_02[..., :3, -1:] += tvec_02
-    trans_02[..., -1, -1:] += 1.0
+    # Instead of incrementally building zeros, assign directly for better performance
+    trans_02 = trans_01.new_zeros(trans_01.shape)
+    trans_02[..., :3, :3] = rmat_02
+    trans_02[..., :3, 3:] = tvec_02
+    trans_02[..., 3, 3] = 1.0
     return trans_02