<!-- CODEFLASH_OPTIMIZATION: {"function":"compose_transformations","file":"kornia/geometry/linalg.py","speedup_pct":"24%","speedup_x":"0.24x","original_runtime":"1.95 milliseconds","best_runtime":"1.58 milliseconds","optimization_type":"memory","timestamp":"2025-08-07T21:40:20.546Z","version":"1.0"} -->
### 📄 24% (0.24x) speedup for ***`compose_transformations` in `kornia/geometry/linalg.py`***

⏱️ Runtime :   **`1.95 milliseconds`**  **→** **`1.58 milliseconds`** (best of `100` runs)
### 📝 Explanation and details

Here is an optimized version of your program. The main optimizations are.

- **Remove Redundant Redefinition:** The local re-definition of `KORNIA_CHECK_IS_TENSOR` is unneeded since you're already importing it, saving namespace lookup and function definition time.
- **Avoid += Where Not Needed:** Replace `+=` with slicing assignment (e.g., `trans_02[..., :3, 0:3] = rmat_02`) to avoid unnecessary initial zeroing and then adding (as `zeros_like` initializes zeros anyway).
- **Reduce Function Calls:** Minimize repeated checks and only perform what is necessary for correctness.
- **Use Fused Tensor Construction:** Instead of calling `zeros_like` and then incrementally building the output, construct the output directly with minimal copies.

Optimized code.



**Key speed improvements:**
- Removed the unnecessary `+=` (which does not benefit from memory aliasing with zeros).
- Removed the duplicate `KORNIA_CHECK_IS_TENSOR` implementation.
- Used `new_zeros` to match dtype/device as input.
- Only performed required assignments to output tensor—no in-place addition needed.
- Used `:3, 3:` for last column assignment (consistent with how slices work; slightly more direct and efficient than `-1:` for a single element).

**Functionality is identical, but now runtime and memory efficiency are improved.**


✅ **Correctness verification report:**

| Test                        | Status            |
| --------------------------- | ----------------- |
| ⚙️ Existing Unit Tests | ✅ **36 Passed** |
| 🌀 Generated Regression Tests | ✅ **40 Passed** |
| ⏪ Replay Tests | 🔘 **None Found** |
| 🔎 Concolic Coverage Tests | 🔘 **None Found** |
|📊 Tests Coverage       | 100.0% |
<details>
<summary>⚙️ Existing Unit Tests and Runtime</summary>

| Test File::Test Function                                                              | Original ⏱️   | Optimized ⏱️   | Speedup   |
|:--------------------------------------------------------------------------------------|:--------------|:---------------|:----------|
| `geometry/test_linalg.py::TestComposeTransforms.test_exception`                       | 7.58μs        | 7.58μs         | ✅0.000%  |
| `geometry/test_linalg.py::TestComposeTransforms.test_smoke`                           | 71.8μs        | 57.6μs         | ✅24.7%   |
| `geometry/test_linalg.py::TestComposeTransforms.test_translation_4x4`                 | 35.2μs        | 27.2μs         | ✅29.0%   |
| `geometry/test_linalg.py::TestComposeTransforms.test_translation_Bx4x4`               | 118μs         | 96.7μs         | ✅22.1%   |
| `geometry/test_linalg.py::TestRelativeTransformation.test_rotation_translation_Bx4x4` | 105μs         | 85.9μs         | ✅22.8%   |
| `geometry/test_linalg.py::TestRelativeTransformation.test_translation_4x4`            | 30.0μs        | 23.0μs         | ✅30.0%   |

</details>

<details>
<summary>🌀 Generated Regression Tests and Runtime</summary>

```python
from __future__ import annotations

from typing import Optional

# imports
import pytest  # used for our unit tests
import torch
from kornia.core import Tensor, zeros_like
from kornia.geometry.linalg import compose_transformations
from typing_extensions import TypeGuard


def zeros_like(x):
    return torch.zeros_like(x)
from kornia.geometry.linalg import compose_transformations

# unit tests

# ----------------- BASIC TEST CASES -----------------

def test_identity_4x4():
    # Compose two identity 4x4 matrices: should return identity
    t1 = torch.eye(4)
    t2 = torch.eye(4)
    codeflash_output = compose_transformations(t1, t2); result = codeflash_output # 35.3μs -> 27.8μs (27.2% faster)

def test_identity_batch():
    # Compose two batches of identity transforms
    t1 = torch.eye(4).repeat(5, 1, 1)
    t2 = torch.eye(4).repeat(5, 1, 1)
    codeflash_output = compose_transformations(t1, t2); result = codeflash_output # 42.0μs -> 35.2μs (19.3% faster)

def test_translation_only():
    # Compose two pure translations
    t1 = torch.eye(4)
    t2 = torch.eye(4)
    t1[:3, 3] = torch.tensor([1., 2., 3.])
    t2[:3, 3] = torch.tensor([4., 5., 6.])
    codeflash_output = compose_transformations(t1, t2); result = codeflash_output # 32.9μs -> 25.2μs (30.2% faster)
    expected = torch.eye(4)
    expected[:3, 3] = torch.tensor([1., 2., 3.]) + torch.tensor([4., 5., 6.])



def test_batch_mixed():
    # Compose batch of transforms with different translations
    batch = 10
    t1 = torch.eye(4).repeat(batch, 1, 1)
    t2 = torch.eye(4).repeat(batch, 1, 1)
    t1[:, 0, 3] = torch.arange(batch).float()
    t2[:, 1, 3] = torch.arange(batch).float()
    codeflash_output = compose_transformations(t1, t2); result = codeflash_output # 49.2μs -> 40.1μs (22.8% faster)
    # Each result[i] should have translation (t1[i,0,3], t2[i,1,3], 0)
    for i in range(batch):
        pass

# ----------------- EDGE TEST CASES -----------------

def test_shape_mismatch_dim():
    # One input is 4x4, one is Nx4x4
    t1 = torch.eye(4)
    t2 = torch.eye(4).repeat(2, 1, 1)
    with pytest.raises(ValueError):
        compose_transformations(t1, t2) # 2.62μs -> 2.67μs (1.54% slower)

def test_shape_mismatch_matrix():
    # Input not 4x4 or Nx4x4
    t1 = torch.eye(3)
    t2 = torch.eye(4)
    with pytest.raises(ValueError):
        compose_transformations(t1, t2) # 2.83μs -> 2.62μs (7.92% faster)

def test_shape_mismatch_last_dim():
    # Input with wrong last dimension
    t1 = torch.zeros(4, 5)
    t2 = torch.zeros(4, 4)
    with pytest.raises(ValueError):
        compose_transformations(t1, t2) # 2.17μs -> 2.12μs (1.93% faster)

def test_not_tensor_input():
    # Inputs are not tensors
    t1 = [[1,0,0,0],[0,1,0,0],[0,0,1,0],[0,0,0,1]]
    t2 = torch.eye(4)
    with pytest.raises(TypeError):
        compose_transformations(t1, t2) # 1.54μs -> 1.58μs (2.59% slower)

def test_nan_inf_values():
    # Inputs contain NaN or Inf
    t1 = torch.eye(4)
    t2 = torch.eye(4)
    t2[0,0] = float('nan')
    codeflash_output = compose_transformations(t1, t2); result = codeflash_output # 37.5μs -> 29.2μs (28.4% faster)
    t2 = torch.eye(4)
    t2[1,1] = float('inf')
    codeflash_output = compose_transformations(t1, t2); result = codeflash_output # 30.9μs -> 23.5μs (31.6% faster)

def test_requires_grad():
    # Inputs with requires_grad = True should preserve property in output
    t1 = torch.eye(4, requires_grad=True)
    t2 = torch.eye(4, requires_grad=True)
    codeflash_output = compose_transformations(t1, t2); result = codeflash_output # 62.0μs -> 45.0μs (37.6% faster)

def test_non_contiguous():
    # Inputs are non-contiguous
    t1 = torch.eye(4).repeat(2, 1, 1)[::2]
    t2 = torch.eye(4).repeat(2, 1, 1)[::2]
    codeflash_output = compose_transformations(t1, t2); result = codeflash_output # 41.6μs -> 34.2μs (21.6% faster)

def test_singleton_batch():
    # Inputs are 1x4x4
    t1 = torch.eye(4).unsqueeze(0)
    t2 = torch.eye(4).unsqueeze(0)
    codeflash_output = compose_transformations(t1, t2); result = codeflash_output # 39.8μs -> 33.1μs (20.4% faster)

def test_broadcast_fail():
    # Inputs with mismatched batch sizes should fail
    t1 = torch.eye(4).repeat(2,1,1)
    t2 = torch.eye(4).repeat(3,1,1)
    with pytest.raises(RuntimeError):
        compose_transformations(t1, t2) # 82.8μs -> 82.1μs (0.862% faster)

# ----------------- LARGE SCALE TEST CASES -----------------

def test_large_batch_identity():
    # Compose large batch of identity transforms
    N = 500  # keep < 1000 for memory
    t1 = torch.eye(4).repeat(N, 1, 1)
    t2 = torch.eye(4).repeat(N, 1, 1)
    codeflash_output = compose_transformations(t1, t2); result = codeflash_output # 84.5μs -> 68.2μs (24.0% faster)


def test_large_batch_translation():
    # Compose large batch of translations only
    N = 999
    t1 = torch.eye(4).repeat(N,1,1)
    t2 = torch.eye(4).repeat(N,1,1)
    t1[:,:3,3] = torch.arange(N).unsqueeze(1).float().repeat(1,3)
    t2[:,:3,3] = torch.arange(N,0,-1).unsqueeze(1).float().repeat(1,3)
    codeflash_output = compose_transformations(t1, t2); result = codeflash_output # 129μs -> 103μs (24.6% faster)
    # Each translation should be t1 + t2
    expected = t1.clone()
    expected[:,:3,3] = t1[:,:3,3] + t2[:,:3,3]

def test_large_batch_requires_grad():
    # Large batch with requires_grad
    N = 123
    t1 = torch.eye(4, requires_grad=True).repeat(N,1,1)
    t2 = torch.eye(4, requires_grad=True).repeat(N,1,1)
    codeflash_output = compose_transformations(t1, t2); result = codeflash_output # 81.3μs -> 62.2μs (30.8% faster)

def test_large_batch_non_contiguous():
    # Large batch, non-contiguous
    N = 100
    t1 = torch.eye(4).repeat(N,1,1)[::2]
    t2 = torch.eye(4).repeat(N,1,1)[::2]
    codeflash_output = compose_transformations(t1, t2); result = codeflash_output # 46.0μs -> 36.5μs (25.9% faster)
# codeflash_output is used to check that the output of the original code is the same as that of the optimized code.

from __future__ import annotations

from typing import Optional

# imports
import pytest  # used for our unit tests
import torch
from kornia.core import Tensor, zeros_like
from kornia.geometry.linalg import compose_transformations
from typing_extensions import TypeGuard

# unit tests

# -------------------------
# Basic Test Cases
# -------------------------

def test_identity_composition_single():
    # Compose two identity 4x4 matrices, result should be identity
    t1 = torch.eye(4)
    t2 = torch.eye(4)
    codeflash_output = compose_transformations(t1, t2); out = codeflash_output # 34.5μs -> 26.8μs (28.6% faster)

def test_identity_composition_batch():
    # Compose two batches of identity matrices
    t1 = torch.eye(4).repeat(5, 1, 1)
    t2 = torch.eye(4).repeat(5, 1, 1)
    codeflash_output = compose_transformations(t1, t2); out = codeflash_output # 40.3μs -> 32.5μs (23.9% faster)

def test_translation_composition():
    # Compose two pure translations
    t1 = torch.eye(4)
    t1[:3, 3] = torch.tensor([1.0, 2.0, 3.0])
    t2 = torch.eye(4)
    t2[:3, 3] = torch.tensor([-1.0, 4.0, 0.5])
    codeflash_output = compose_transformations(t1, t2); out = codeflash_output # 32.8μs -> 25.2μs (29.9% faster)
    # The composed translation should be t1[:3,3] + t2[:3,3]
    expected = torch.eye(4)
    expected[:3, 3] = torch.tensor([1.0, 2.0, 3.0]) + torch.tensor([-1.0, 4.0, 0.5])


def test_mixed_batch_and_single():
    # Compose a batch with a single transformation
    t1 = torch.eye(4).repeat(3, 1, 1)
    t2 = torch.eye(4)
    # This should fail due to mismatched dims
    with pytest.raises(ValueError):
        compose_transformations(t1, t2) # 2.79μs -> 2.92μs (4.25% slower)

def test_mixed_types():
    # Compose float32 and float64 tensors
    t1 = torch.eye(4, dtype=torch.float32)
    t2 = torch.eye(4, dtype=torch.float64)
    # Should work since torch promotes types
    codeflash_output = compose_transformations(t1.to(torch.float64), t2); out = codeflash_output # 40.8μs -> 34.0μs (19.7% faster)

def test_batch_rotation_and_translation():
    # Compose a batch of different transforms
    t1 = torch.eye(4).repeat(2, 1, 1)
    t1[0, :3, 3] = torch.tensor([1.0, 0.0, 0.0])
    t1[1, :3, 3] = torch.tensor([0.0, 2.0, 0.0])
    t2 = torch.eye(4).repeat(2, 1, 1)
    t2[0, :3, 3] = torch.tensor([0.0, 1.0, 0.0])
    t2[1, :3, 3] = torch.tensor([0.0, 0.0, 3.0])
    codeflash_output = compose_transformations(t1, t2); out = codeflash_output # 42.7μs -> 34.7μs (22.9% faster)

# -------------------------
# Edge Test Cases
# -------------------------

def test_non_tensor_input():
    # Inputs that are not tensors should raise TypeError
    with pytest.raises(TypeError):
        compose_transformations([[1, 0, 0, 0],[0,1,0,0],[0,0,1,0],[0,0,0,1]], torch.eye(4)) # 1.62μs -> 1.62μs (0.000% faster)
    with pytest.raises(TypeError):
        compose_transformations(torch.eye(4), None) # 1.25μs -> 1.29μs (3.25% slower)

def test_wrong_shape():
    # Inputs with wrong shape should raise ValueError
    with pytest.raises(ValueError):
        compose_transformations(torch.eye(3), torch.eye(4)) # 2.50μs -> 2.38μs (5.26% faster)
    with pytest.raises(ValueError):
        compose_transformations(torch.eye(4, 5), torch.eye(4)) # 1.42μs -> 1.38μs (2.98% faster)
    with pytest.raises(ValueError):
        compose_transformations(torch.eye(4), torch.ones(5,4,4,4)) # 1.88μs -> 1.96μs (4.24% slower)


def test_non_square_matrix():
    # Inputs that are not square matrices should raise ValueError
    with pytest.raises(ValueError):
        compose_transformations(torch.ones(4,5), torch.eye(4)) # 2.83μs -> 2.88μs (1.46% slower)
    with pytest.raises(ValueError):
        compose_transformations(torch.eye(4), torch.ones(4,5)) # 1.88μs -> 1.83μs (2.24% faster)

def test_last_row_not_homogeneous():
    # Compose two matrices where last row is not [0,0,0,1]
    t1 = torch.eye(4)
    t2 = torch.eye(4)
    t1[-1, :] = torch.tensor([0.0, 0.0, 0.0, 2.0])
    codeflash_output = compose_transformations(t1, t2); out = codeflash_output # 39.2μs -> 30.0μs (30.4% faster)

def test_empty_batch():
    # Compose empty batch
    t1 = torch.empty(0, 4, 4)
    t2 = torch.empty(0, 4, 4)
    codeflash_output = compose_transformations(t1, t2); out = codeflash_output # 40.8μs -> 32.5μs (25.5% faster)

def test_requires_grad():
    # Compose tensors with requires_grad=True
    t1 = torch.eye(4, requires_grad=True)
    t2 = torch.eye(4, requires_grad=True)
    codeflash_output = compose_transformations(t1, t2); out = codeflash_output # 63.9μs -> 46.9μs (36.1% faster)

# -------------------------
# Large Scale Test Cases
# -------------------------

def test_large_batch():
    # Compose two large batches (size 999 to stay under 100MB)
    N = 999
    t1 = torch.eye(4).repeat(N, 1, 1)
    t2 = torch.eye(4).repeat(N, 1, 1)
    codeflash_output = compose_transformations(t1, t2); out = codeflash_output # 128μs -> 102μs (25.7% faster)

def test_large_random_batch():
    # Compose two large batches of random transformations
    N = 500
    # Random rotation matrices using QR decomposition
    rand = torch.randn(N, 3, 3)
    q, _ = torch.linalg.qr(rand)
    t1 = torch.eye(4).repeat(N, 1, 1)
    t2 = torch.eye(4).repeat(N, 1, 1)
    t1[:, :3, :3] = q
    t1[:, :3, 3] = torch.randn(N, 3)
    t2[:, :3, :3] = q.transpose(1,2)  # inverse rotation
    t2[:, :3, 3] = -torch.randn(N, 3)
    codeflash_output = compose_transformations(t1, t2); out = codeflash_output # 82.6μs -> 66.5μs (24.3% faster)

def test_large_scale_memory():
    # Test that function does not crash or hang on large input
    N = 800
    t1 = torch.eye(4).repeat(N, 1, 1)
    t2 = torch.eye(4).repeat(N, 1, 1)
    codeflash_output = compose_transformations(t1, t2); out = codeflash_output # 106μs -> 85.8μs (24.3% faster)

def test_large_scale_noncontiguous():
    # Test with non-contiguous tensors
    N = 100
    t1 = torch.eye(4).repeat(N, 1, 1)[::2]
    t2 = torch.eye(4).repeat(N, 1, 1)[::2]
    codeflash_output = compose_transformations(t1, t2); out = codeflash_output # 45.1μs -> 35.6μs (26.7% faster)

# -------------------------
# Regression/Mutation Tests
# -------------------------

def test_output_is_new_tensor():
    # Output must not be a view of the input
    t1 = torch.eye(4)
    t2 = torch.eye(4)
    codeflash_output = compose_transformations(t1, t2); out = codeflash_output # 35.0μs -> 27.3μs (27.9% faster)
    out[0,0] = 123.456

def test_no_inplace_on_inputs():
    # Inputs must not be modified
    t1 = torch.eye(4)
    t2 = torch.eye(4)
    t1_clone = t1.clone()
    t2_clone = t2.clone()
    codeflash_output = compose_transformations(t1, t2); _ = codeflash_output # 33.7μs -> 25.9μs (30.1% faster)
# codeflash_output is used to check that the output of the original code is the same as that of the optimized code.
```

</details>